Add files via upload
This script is relatively straightforward. MATLAB offers built-in functions that make solving this particular problem seem trivial. Those functions are factor() and max(). I used the first to create a vector with all of the factors of the number, and the second function was used to find the largest element of the function.