Add files via upload
This program uses nested loops to find the largest palindrome that is a product of two n digit numbers. It is interesting to note, that a slightly faster version of this program can be written using the built in MATLAB function, meshgrid instead of nested loops. I will upload this soon.